+++ source/battle-board/battle-board/models/interfaces/user.ts
